Move committee directives to common
